fix(fake-driver): create cjs wrapper
