+++ packages/fake-driver/test/driver-e2e-specs.js
@@ -2,7 +2,7 @@
 
 import axios from 'axios';
 import { baseDriverE2ETests } from '@appium/base-driver/build/test/basedriver';
-import { FakeDriver, startServer } from '../index.js';
+import { FakeDriver, startServer } from '../lib/index.js';
 import { BASE_CAPS, deleteSession, initSession, TEST_HOST, TEST_PORT, W3C_PREFIXED_CAPS } from './helpers';
 import contextTests from './context-tests';
 import findElementTests from './find-element-tests';
