+++ packages/fake-driver/test/driver-specs.js
@@ -1,7 +1,7 @@
 // transpile:mocha
 
 import _ from 'lodash';
-import { FakeDriver } from '..';
+import { FakeDriver } from '../lib';
 import { W3C_CAPS, W3C_PREFIXED_CAPS } from './helpers';
 import { baseDriverUnitTests } from '@appium/base-driver/build/test/basedriver';
